Struct rustc_interface::interface::Config

pub struct Config {

    pub opts: Options,
    pub crate_cfg: FxHashSet<(String, Option<String>)>,
    pub crate_check_cfg: CheckCfg,
    pub input: Input,
    pub output_dir: Option<PathBuf>,
    pub output_file: Option<OutFileName>,
    pub ice_file: Option<PathBuf>,
    pub file_loader: Option<Box<dyn FileLoader + Send + Sync>>,
    pub locale_resources: &'static [&'static str],
    pub lint_caps: FxHashMap<LintId, Level>,
    pub parse_sess_created: Option<Box<dyn FnOnce(&mut ParseSess) + Send>>,
    pub register_lints: Option<Box<dyn Fn(&Session, &mut LintStore) + Send + Sync>>,
    pub override_queries: Option<fn(_: &Session, _: &mut Providers, _: &mut ExternProviders)>,
    pub make_codegen_backend: Option<Box<dyn FnOnce(&Options) -> Box<dyn CodegenBackend> + Send>>,
    pub registry: Registry,

}

The compiler configuration

Fields
opts: Options

Command line options

crate_cfg: FxHashSet<(String, Option<String>)>

cfg! configuration in addition to the default ones

crate_check_cfg: CheckCfginput: Inputoutput_dir: Option<PathBuf>output_file: Option<OutFileName>ice_file: Option<PathBuf>file_loader: Option<Box<dyn FileLoader + Send + Sync>>locale_resources: &'static [&'static str]lint_caps: FxHashMap<LintId, Level>parse_sess_created: Option<Box<dyn FnOnce(&mut ParseSess) + Send>>

This is a callback from the driver that is called when ParseSess is created.

register_lints: Option<Box<dyn Fn(&Session, &mut LintStore) + Send + Sync>>

This is a callback from the driver that is called when we’re registering lints; it is called during plugin registration when we have the LintStore in a non-shared state.

Note that if you find a Some here you probably want to call that function in the new function being registered.

override_queries: Option<fn(_: &Session, _: &mut Providers, _: &mut ExternProviders)>

This is a callback from the driver that is called just after we have populated the list of queries.

The second parameter is local providers and the third parameter is external providers.

make_codegen_backend: Option<Box<dyn FnOnce(&Options) -> Box<dyn CodegenBackend> + Send>>

This is a callback from the driver that is called to create a codegen backend.

registry: Registry

Registry of diagnostics codes.