rustc_interface

Module limits

Registering limits:

recursion_limit: there are various parts of the compiler that must impose arbitrary limits on how deeply they recurse to prevent stack overflow.

move_size_limit

type_length_limit

pattern_complexity_limit

Users can override these limits via an attribute on the crate like #![recursion_limit="22"]. This pass just looks for those attributes.

Functions

get_limit 🔒

get_recursion_limit 🔒

provide 🔒