rustc_interface::util
Function get_codegen_backend

pub fn get_codegen_backend(
    early_dcx: &EarlyDiagCtxt,
    sysroot: &Path,
    backend_name: Option<&str>,
    target: &Target,
) -> Box<dyn CodegenBackend>

Get the codegen backend based on the name and specified sysroot.

A name of None indicates that the default backend should be used.