Crate rustc_interface

Re-exports

pub use interface::Config;

pub use interface::run_compiler;

pub use passes::DEFAULT_QUERY_PROVIDERS;

pub use passes::create_and_enter_global_ctxt;

pub use passes::parse;

Modules

callbacks 🔒

Throughout the compiler tree, there are several places which want to have access to state or queries while being inside crates that are dependencies of rustc_middle. To facilitate this, we have the rustc_data_structures::AtomicRef type, which allows us to setup a global static which can then be set in this file at program startup.

errors

interface

limits 🔒

Registering limits:

passes

proc_macro_decls 🔒

queries 🔒

util

Structs

Linker

Statics

DEFAULT_LOCALE_RESOURCE

Raw content of Fluent resource for this crate, generated by fluent_messages macro, imported by rustc_driver to include all crates’ resources in one bundle.

Functions

setup_callbacks

Sets up the callbacks in prior crates which we want to refer to the TyCtxt in.