rustc_interface
Module interface

Structs

Compiler

Represents a compiler session. Note that every Compiler contains a Session, but Compiler also contains some things that cannot be in Session, due to Session being in a crate that has many fewer dependencies than this crate.

Config

The compiler configuration

Functions

initialize_checked_jobserver 🔒

Initialize jobserver before getting jobserver::client and build_session.

parse_cfg 🔒

Converts strings provided as --cfg [cfgspec] into a Cfg.

parse_check_cfg 🔒

Converts strings provided as --check-cfg [specs] into a CheckCfg.

run_compiler

try_print_query_stack

Type Aliases

Result