Struct rustc_interface::interface::Compiler

pub struct Compiler {
    pub(crate) sess: Lrc<Session>,
    codegen_backend: Lrc<dyn CodegenBackend>,
    pub(crate) register_lints: Option<Box<dyn Fn(&Session, &mut LintStore) + Send + Sync>>,
    pub(crate) override_queries: Option<fn(_: &Session, _: &mut Providers, _: &mut ExternProviders)>,
}

Represents a compiler session. Note that every Compiler contains a Session, but Compiler also contains some things that cannot be in Session, due to Session being in a crate that has many fewer dependencies than this crate.

Can be used to run rustc_interface queries. Created by passing Config to run_compiler.

Fields
sess: Lrc<Session>codegen_backend: Lrc<dyn CodegenBackend>register_lints: Option<Box<dyn Fn(&Session, &mut LintStore) + Send + Sync>>override_queries: Option<fn(_: &Session, _: &mut Providers, _: &mut ExternProviders)>
Implementations

impl Compiler

pub fn session(&self) -> &Lrc<Session>

pub fn codegen_backend(&self) -> &Lrc<dyn CodegenBackend>

pub fn register_lints( &self ) -> &Option<Box<dyn Fn(&Session, &mut LintStore) + Send + Sync>>

pub fn build_output_filenames( &self, sess: &Session, attrs: &[Attribute] ) -> OutputFilenames

impl Compiler

pub fn enter<F, T>(&self, f: F) -> Twhere F: for<'tcx> FnOnce(&'tcx Queries<'tcx>) -> T,